+++ changelog.md
@@ -44,6 +44,7 @@ This version is not yet released. If you are reading this on the website, then t
   - Each additional argument increases the number of arguments passed to [`reduce /`](https://uiua.org/docs/reduce) by 1
   - Additional arguments are passed to the function on every iteration
 - [`try ⍣`](https://uiua.org/docs/try) signature checking is now more permissive
+- Switch function signature checking is now more permissive with branches that have an [`assert ⍤`](https://uiua.org/docs/assert) that always triggers
 - Add the [`&exit`](https://uiua.org/docs/&exit) system function, which exits the program with a status code
 - Add the experimental [`&memcpy`](https://uiua.org/docs/&memcpy) system function, which copies the data from [`&ffi`](https://uiua.org/docs/&ffi) pointers to an array
 - Add the experimental [`&memfree`](https://uiua.org/docs/&memfree) system function, which frees memory allocated by [`&ffi`](https://uiua.org/docs/&ffi) functions